+++ Kernel/Modules/CustomerTicketZoom.pm
@@ -1375,9 +1375,11 @@ sub _Mask {
         $LayoutObject->Block(
             Name => 'TicketDynamicField',
             Data => {
-                Label => $Label,
-                Value => $ValueStrg->{Value},
-                Title => $ValueStrg->{Title},
+                Label       => $Label,
+                Value       => $ValueStrg->{Value},
+                Title       => $ValueStrg->{Title},
+                Link        => $DynamicFieldConfig->{Config}->{Link},
+                LinkPreview => $DynamicFieldConfig->{Config}->{LinkPreview},
             },
         );